Move genesis generate
